requirements.txt
google-generativeai>=0.5.2
fastapi>=0.112.0
uvicorn>=0.30.0
pydantic>=2.7.0
google-adk>=1.18.0
python-dotenv